lots of small changes to get these objects to use the Max5/6 API.  lots of them were using class_new() but then calling newobject() or forgetting to call class_register(), etc.

+++ src/2threshattack~/2threshattack~.c
@@ -59,8 +59,8 @@ typedef struct _tta
     t_pxobject x_obj;	//  header
     void *event_outlet;
     
-    float low;
-    float high;
+    double low;
+    double high;
     int in_attack;
 } t_tta;
